Refactor build code to libraries
